Restructure modules for GCloud integration.
